requirements.txt
numpy == 1.24.2
pandas == 1.4.4
xlwt == 1.3.0
PyMuPDF == 1.22.5
fitz == 0.0.1.dev2
datetime == 5.1
openpyxl == 3.0.10
easyocr == 1.7.0
pillow == 9.2.0
opencv-python == 4.8.0.74
watchdog == 2.1.6




